requirements.txt
aioredis==2.0.1
aiosmtplib==1.1.6
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.1
async-timeout==4.0.2
attrs==21.4.0
bcrypt==3.2.2
black==22.3.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
Deprecated==1.2.13
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
fakeredis==1.7.5
fastapi==0.75.1
fastapi-mail==1.0.8
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
iniconfig==1.1.1
isort==5.10.1
Jinja2==3.1.2
Mako==1.2.0
mangum==0.15.0
MarkupSafe==2.1.1
mypy-extensions==0.4.3
packaging==21.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
pyparsing==3.0.8
pytest==7.1.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
redis==4.3.1
requests==2.27.1
responses==0.20.0
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.36
starlette==0.17.1
tomli==2.0.1
typing-extensions==4.2.0
urllib3==1.26.9
uvicorn==0.15.0
wrapt==1.14.1
zipp==3.8.0
